+++ tests/unit/work/_fake_connection.py
@@ -25,7 +25,7 @@ from unittest import mock
 import pytest
 
 from neo4j import ServerInfo
-from neo4j._io.deadline import Deadline
+from neo4j._deadline import Deadline
 
 
 class FakeConnection(mock.NonCallableMagicMock):
